+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Markup of time tag used on automatic updates was fixed.
+  [hvelarde]
 
 
 1.0a1 (2014-09-01)